+++ umc/src/main/java/com/umc/domain/comment/entity/Comment.java
@@ -1,9 +1,9 @@
 package com.umc.domain.comment.entity;
 
 import com.umc.common.entity.BaseTimeEntity;
+
 import javax.persistence.*;
 
-import com.umc.domain.board.entity.Board;
 import com.umc.domain.post.entity.Post;
 import com.umc.domain.user.entity.Member;
 import lombok.*;
@@ -18,7 +18,7 @@ public class Comment extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "post_id")
